backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.2
langdetect==1.0.9
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
more-itertools==10.8.0
google-generativeai==0.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pymongo==4.6.0
pillow==10.1.0
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
